requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.0
sentence-transformers==2.2.2
tidb-vector==0.0.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
requests==2.31.0
